Fix pinboard tags response

+++ app/src/main/kotlin/com/fibelatti/pinboard/features/tags/data/TagsDataSource.kt
@@ -50,7 +50,7 @@ class TagsDataSource @Inject constructor(
         tagsApi.getTags()
     }.mapCatching { tagsAndPostCount ->
         tagsAndPostCount
-            .map { (tag, postCount) -> Tag(tag, postCount.toInt()) }
+            .map { (tag, postCount) -> Tag(tag, postCount) }
             .sortedBy { it.name }
     }
